Ships Refactored
Ship by default has a base place and attack function along with bound checks. Abstracted the captains module functionality, and 3 children that utilize the captains quarters
